:sparkles: ユーザー名・アイコンの変更を実装
