+++ src/layout/AuthLayout.tsx
@@ -1,7 +1,7 @@
 import { Auth, Button, IconLogOut } from "@supabase/ui";
 import type { CustomLayout } from "next";
-import { cloneElement, useCallback, useEffect, useState } from "react";
-import { addNewProfile, client, getProfile } from "src/lib/SupabaseClient";
+import { useCallback, useEffect, useState } from "react";
+import { client } from "src/lib/SupabaseClient";
 
 import { Footer } from "./Footer";
 import { Header } from "./Header";
@@ -18,33 +18,12 @@ export const AuthLayout: CustomLayout = (props: Props) => {
   const { user } = Auth.useUser();
 
   const [isMounted, setIsMounted] = useState<boolean>(false);
-  const [username, setUsername] = useState<string>("");
-  const [avatar, setAvatar] = useState<string>("");
 
   const { children } = props;
-
-  const fetchProfile = useCallback(async (uid: string) => {
-    const profile = await getProfile();
-    if (profile) {
-      setUsername(profile.username);
-      setAvatar(profile.avatar);
-    } else {
-      const isOk = await addNewProfile(uid);
-      if (!isOk) {
-        alert("プロフィールの新規登録に失敗しました。");
-      }
-    }
-  }, []);
 
   const handleLogout = useCallback(() => {
     client.auth.signOut();
   }, []);
-
-  useEffect(() => {
-    if (user) {
-      fetchProfile(user.id);
-    }
-  }, [user, fetchProfile]);
 
   useEffect(() => {
     setIsMounted(true);
@@ -53,20 +32,13 @@ export const AuthLayout: CustomLayout = (props: Props) => {
   return (
     <div className="flex flex-col min-h-screen">
       <header>
-        <Header avatar={avatar} />
+        <Header />
       </header>
       <main className="flex-1 px-4">
         <LayoutErrorBoundary>
           {isMounted && user ? (
             <div>
-              <div>
-                {cloneElement(children, {
-                  uid: user.id,
-                  username: username,
-                  avatar: avatar,
-                  fetchProfile: fetchProfile,
-                })}
-              </div>
+              <div>{children}</div>
               <div className="flex justify-end my-4 mx-2">
                 <Button
                   size="medium"
